refactor(web): split media utils into labels and normalize modules
media.ts mixed enum-to-display mapping with API-shape reshaping. labels.ts
now holds the per-type presentation config, normalize.ts the view-model
mapping; media.ts stays as a barrel so imports elsewhere are unchanged.
